Added block in scanner Refactored the name of Factory Added Safeguard for entry value Changed object menu to string menu
