+++ src/main/java/Main.java
@@ -1,4 +1,4 @@
-import basket.factory.Factory;
+import basket.factory.FactoryMenageShopping;
 import menu.Menu;
 
 
@@ -6,8 +6,8 @@ public class Main {
     public static void main(String[] args) {
         System.out.println("Welcome. Please choose number " +
                 "from menu below:");
-        new Menu().menu();
-        new Factory().chooseNumberFromMenu();
+        new Menu().MenuPrinter();
+        new FactoryMenageShopping().chooseNumberFromMenu();
 
         System.out.println("End of shopping. See you again");
     }
